1. 恢复多余的修改 2. 增加示例代码 examples/fastapi_stream.py 3. requirements增加fastapi
