+++ examples/ping.py
@@ -1,43 +1,26 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+@Time    : 2024/4/22 14:28
+@Author  : alexanderwu
+@File    : ping.py
+"""
+
 import asyncio
-import threading
-import queue as sync_queue
-from fastapi import FastAPI
-from sse_starlette.sse import EventSourceResponse
-import uvicorn
+
 from metagpt.llm import LLM
+from metagpt.logs import logger
 
-from metagpt.roles.di.data_interpreter import DataInterpreter
 
-app = FastAPI()
-llm = LLM()
-async def generate_async_stream(queue):
-    requirement = "解决这个数学问题：正整数m和n的最大公约数是6。m和n的最小公倍数是126。m + n的最小可能值是多少？"
-    await llm.aask(requirement, stream=True, queue=queue)
+async def ask_and_print(question: str, llm: LLM, system_prompt) -> str:
+    logger.info(f"Q: {question}")
+    rsp = await llm.aask(question, system_msgs=[system_prompt])
+    logger.info(f"A: {rsp}")
+    logger.info("\n")
+    return rsp
 
-@app.get("/stream")
-def stream():
-    queue = sync_queue.Queue()
-    print(queue)
 
-    def run_loop(queue):
-        asyncio.set_event_loop(asyncio.new_event_loop())
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(generate_async_stream(queue))
+async def main():
+    llm = LLM()
+    await ask_and_print("ping?", llm, "Just answer pong when ping.")
 
-    thread = threading.Thread(target=run_loop, args=(queue,))
-    thread.start()
-
-    def generate_sync_stream():
-        while True:
-            message = queue.get()
-            print(message)
-            if message is None:
-                break
-            yield f"data: {message}\n\n"
-
-    return EventSourceResponse(generate_sync_stream(), media_type='text/event-stream')
-
-if __name__ == "__main__":
-    uvicorn.run(app='examples.ping:app',
-                host="0.0.0.0",
-                port=3000)
